voice rewrite first working state
